# [1.9] AssetKey path as a tuple