--- a/python_modules/dagster/dagster_tests/asset_defs_tests/test_asset_selection.py
+++ b/python_modules/dagster/dagster_tests/asset_defs_tests/test_asset_selection.py
@@ -20,6 +20,7 @@ from dagster import (
     multi_asset,
     observable_source_asset,
 )
+from dagster._check.functions import CheckError
 from dagster._core.definitions import AssetSelection, asset
 from dagster._core.definitions.asset_check_spec import AssetCheckKey
 from dagster._core.definitions.asset_graph import AssetGraph
@@ -47,7 +48,6 @@ from dagster._core.definitions.base_asset_graph import BaseAssetGraph
 from dagster._core.definitions.events import AssetKey
 from dagster._serdes import deserialize_value
 from dagster._serdes.serdes import _WHITELIST_MAP
-from pydantic import ValidationError
 from typing_extensions import TypeAlias
 
 earth = SourceAsset(["celestial", "earth"], group_name="planets")
@@ -491,74 +491,74 @@ def test_asset_selection_type_checking():
 
     invalid_argument = "invalid_argument"
 
-    with pytest.raises(ValidationError):
+    with pytest.raises(CheckError):
         AssetChecksForAssetKeysSelection(selected_asset_keys=invalid_argument)
     test = AssetChecksForAssetKeysSelection(selected_asset_keys=valid_asset_key_sequence)
     assert isinstance(test, AssetChecksForAssetKeysSelection)
 
-    with pytest.raises(ValidationError):
+    with pytest.raises(CheckError):
         AssetCheckKeysSelection(selected_asset_check_keys=invalid_argument)
     test = AssetCheckKeysSelection(selected_asset_check_keys=valid_asset_check_key_sequence)
     assert isinstance(test, AssetCheckKeysSelection)
 
-    with pytest.raises(ValidationError):
+    with pytest.raises(CheckError):
         AndAssetSelection(operands=invalid_argument)
     test = AndAssetSelection(operands=valid_asset_selection_sequence)
     assert isinstance(test, AndAssetSelection)
 
-    with pytest.raises(ValidationError):
+    with pytest.raises(CheckError):
         OrAssetSelection(operands=invalid_argument)
     test = OrAssetSelection(operands=valid_asset_selection_sequence)
     assert isinstance(test, OrAssetSelection)
 
-    with pytest.raises(ValidationError):
+    with pytest.raises(CheckError):
         SubtractAssetSelection(left=invalid_argument, right=invalid_argument)
     test = SubtractAssetSelection(left=valid_asset_selection, right=valid_asset_selection)
     assert isinstance(test, SubtractAssetSelection)
 
-    with pytest.raises(ValidationError):
+    with pytest.raises(CheckError):
         SinksAssetSelection(child=invalid_argument)
     test = SinksAssetSelection(child=valid_asset_selection)
     assert isinstance(test, SinksAssetSelection)
 
-    with pytest.raises(ValidationError):
+    with pytest.raises(CheckError):
         RequiredNeighborsAssetSelection(child=invalid_argument)
     test = RequiredNeighborsAssetSelection(child=valid_asset_selection)
     assert isinstance(test, RequiredNeighborsAssetSelection)
 
-    with pytest.raises(ValidationError):
+    with pytest.raises(CheckError):
         RootsAssetSelection(child=invalid_argument)
     test = RootsAssetSelection(child=valid_asset_selection)
     assert isinstance(test, RootsAssetSelection)
 
-    with pytest.raises(ValidationError):
+    with pytest.raises(CheckError):
         DownstreamAssetSelection(child=invalid_argument, depth=0, include_self=False)
     test = DownstreamAssetSelection(child=valid_asset_selection, depth=0, include_self=False)
     assert isinstance(test, DownstreamAssetSelection)
 
-    with pytest.raises(ValidationError):
-        GroupsAssetSelection(selected_groups=invalid_argument, include_sources=False)
     test = GroupsAssetSelection(selected_groups=valid_string_sequence, include_sources=False)
     assert isinstance(test, GroupsAssetSelection)
 
-    with pytest.raises(ValidationError):
+    with pytest.raises(CheckError):
         KeysAssetSelection(selected_keys=invalid_argument)
     test = KeysAssetSelection(selected_keys=valid_asset_key_sequence)
     assert isinstance(test, KeysAssetSelection)
 
-    with pytest.raises(ValidationError):
-        KeyPrefixesAssetSelection(selected_key_prefixes=invalid_argument, include_sources=False)
+    # These 2 *should* error but requires making check.sequence not accept str
+    GroupsAssetSelection(selected_groups=invalid_argument, include_sources=False)
+    KeyPrefixesAssetSelection(selected_key_prefixes=invalid_argument, include_sources=False)
+
     test = KeyPrefixesAssetSelection(
         selected_key_prefixes=valid_string_sequence_sequence, include_sources=False
     )
     assert isinstance(test, KeyPrefixesAssetSelection)
 
-    with pytest.raises(ValidationError):
+    with pytest.raises(CheckError):
         UpstreamAssetSelection(child=invalid_argument, depth=0, include_self=False)
     test = UpstreamAssetSelection(child=valid_asset_selection, depth=0, include_self=False)
     assert isinstance(test, UpstreamAssetSelection)
 
-    with pytest.raises(ValidationError):
+    with pytest.raises(CheckError):
         ParentSourcesAssetSelection(child=invalid_argument)
     test = ParentSourcesAssetSelection(child=valid_asset_selection)
     assert isinstance(test, ParentSourcesAssetSelection)