@alangenfeld 2024-10-14
The only real consequence of this change that i can point to is comparing `key.path` with a list no longer works. The majority of changes in this PR are fixing those call-sites. Makes sense that happen in our test code - seems far less likely to happen in user code but theres some risk there.