@schrockn 2024-10-14
To game out the totally safe version:

* Change `path` to a `cached_property` that returns a list
* Change storage to `path_tuple` or something and do `whitelist_for_serdes` stuff to deal with compat

I am missing anything?