@alangenfeld 2024-10-14
good call - i think it might just be https://github.com/dagster-io/dagster/pull/25261/files will see if tests pass