@alangenfeld 2024-10-15
some example callsites that were doing `asset_key[0]` instead of `asset_key.path` https://github.com/dagster-io/internal/pull/12181/files