@schrockn 2024-10-15
Despite technically being a breaking change I think it is should be viewed as a bug fix. The examples you had to change are pretty damning. If this goes smoothly I think we should consider banning `NamedTuple` from public APIs and moving all to `@record`.